basic tab functionality in settings modal
